Rename test to tests
