+++ tests/test_consumer.py
@@ -1,8 +1,6 @@
-import sys
-from test import conftest
 from unittest.mock import MagicMock, patch
 
-import pytest
+from tests import conftest
 
 mock_confluent_avro_consumer = conftest.mock_confluent_avro_consumer
